requirements.txt
alembic==1.14.0
amqp==5.3.1
anyio==4.8.0
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
cache==1.0.3
cachelib==0.13.0
celery==5.3.6
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
codeviz==0.1.0
colorama==0.4.6
ConfigArgParse==1.7
contourpy==1.3.2
cryptography==41.0.7
cycler==0.12.1
dnspython==2.7.0
dominate==2.9.1
email_validator==2.2.0
Flask==2.2.5
Flask-Admin==1.6.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.3.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Moment==1.0.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.57.0
gevent==24.11.1
geventhttpclient==2.3.3
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.4.8
kombu==5.4.2
locust==2.34.1
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
msgpack==1.1.0
numpy==2.2.4
packaging==24.2
pillow==11.2.1
prompt_toolkit==3.0.50
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydeps==3.0.1
PyJWT==2.8.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==20.7
pytz==2025.1
pyzmq==26.4.0
redis==5.2.1
requests==2.32.3
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
stdlib-list==0.11.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
vine==5.1.0
visitor==0.1.3
wcwidth==0.2.13
Werkzeug==2.2.3
WTForms==3.2.1
zope.event==5.0
zope.interface==7.2
